CommandListener: Fix compilation in case QSAP_WLAN is undefined
startSoftap has an additional parameter that was not passed
when QSAP_WLAN was not defined.

In case QSAP_WLAN is not defined, do not ask to use global hostapd
control interface since this is relevant only in case FST feature
is enabled which requires QSAP_WLAN to be defined.

Change-Id: Icfda9d77013a3a2e7fe3be0b83ef4e8a494ca61c
CRs-Fixed: 1058571