DO NOT MERGE Fix getaddrinfo test to use random mapping ids
This seems to have been intended but it's just using the loop index. In
fact the only thing the random index is used for is an assert to check
that it is a valid index.

Also stops the test segfaulting if num_queries > num_hosts

(cherry picked from commit c4aba0ab2e9539a5e185e3800d47b3fcda9af779)

Change-Id: Iae2369676c08fa8f8110c3cdebb0e4294a60b4b2